Split settings up in regular and local settings
Many settings are actually global and required for the app to run
correctly. The sane defaults for these are now in settings.py.

Other settings that change on an app-to-app basis have been moved
to "local_settings.py" ("local_settings.example.py").

Also note that the sunlight labs API key can now be inserted
directly into the local_settings file.